requirements.txt
pydantic>=2.7,<3
pydantic-settings>=2.2,<3
typer>=0.12,<1
httpx>=0.27,<1
feedparser>=6,<7
tenacity>=8.3,<9
sqlalchemy>=2,<3
sqlite-utils>=3.36,<4
orjson>=3.10,<4
google-generativeai>=0.7,<1
pytest>=8,<9
matplotlib>=3.7,<4
pandas>=2.0,<3
